Typecast DoubleValue to Int

+++ test/metric/container_insights_util.go
@@ -330,7 +330,7 @@ func ValidateNeuronCoreUtilizationValuesLogs(env *environment.MetaData) status.T
 			if strings.HasPrefix(coreKey, "core") {
 				coreNumStr := strings.TrimPrefix(coreKey, "core")
 				expectedValue, err := strconv.Atoi(coreNumStr)
-				if err != nil || float64(actualValue) != float64(expectedValue) {
+				if err != nil || int(actualValue) != int(expectedValue) {
 					log.Printf("Core utilization validation failed: expected %s:%d, got %v", 
 						coreKey, expectedValue, actualValue)
 					return testResult
